requirements.txt
asgiref==3.9.1
autoflake==2.3.1
autopep8==2.3.2
bandit==1.8.6
black==25.1.0
cfgv==3.4.0
click==8.2.1
distlib==0.3.9
Django==4.2.17
django-stubs==5.2.1
django-stubs-ext==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==37.4.0
filelock==3.18.0
flake8==7.3.0
identify==2.6.12
iniconfig==2.1.0
isort==6.0.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.16.1
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.1
pytest-django==4.11.1
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
stevedore==5.4.1
tomli==2.2.1
types-PyYAML==6.0.12.20250516
typing_extensions==4.14.1
tzdata==2025.2
virtualenv==20.31.2
